+++ starsky/starskycore/Services/ExifCopy.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using starskycore.Helpers;
 using starskycore.Interfaces;
 using starskycore.Models;
@@ -53,14 +54,29 @@ namespace starskycore.Services
 		private List<string> CompareAll(FileIndexItem fileIndexItem)
 		{
 			return FileIndexCompareHelper.Compare(new FileIndexItem(), fileIndexItem);
-		} 
-	    
-		public void CopyExifToThumbnail(string subPath, string thumbPath)
-		{
-			var updateModel = _readMeta.ReadExifAndXmpFromFile(subPath);
-			var comparedNames = CompareAll(updateModel);
-
-			new ExifToolCmdHelper(_exifTool,_iStorage,_readMeta).Update(updateModel, comparedNames);
 		}
+
+		/// <summary>
+		/// Copy for list all items (ignore if file already exist)
+		/// </summary>
+		/// <param name="fileIndexList">List with all content to be synced</param>
+		public void CopyExifToThumbnail(IEnumerable<FileIndexItem> fileIndexList)
+		{
+			foreach ( var updateModel in fileIndexList )
+			{
+				if ( ! _iStorage.ThumbnailExist(updateModel.FileHash) ) continue;
+				var comparedNames = CompareAll(updateModel);
+				comparedNames.Add(nameof(FileIndexItem.Software));
+				new ExifToolCmdHelper(_exifTool,_iStorage,_readMeta).UpdateThumbnail(updateModel, comparedNames);
+			}
+		}
+	    
+//		public void CopyExifToThumbnail(string subPath, string thumbPath)
+//		{
+//			var updateModel = _readMeta.ReadExifAndXmpFromFile(subPath);
+//			var comparedNames = CompareAll(updateModel);
+//
+//			new ExifToolCmdHelper(_exifTool,_iStorage,_readMeta).Update(updateModel, comparedNames);
+//		}
 	}
 }