+++ history.md
@@ -25,8 +25,8 @@
 - [x]   check: /api/delete
 - [x]   check: /api/download photo
 - []    check: xmpSync for exiftool
-- []   check: starskyWebHtmlCli
-- []   check: ResizeOverlayImage in webHtmlCli
+- [x]   check: starskyWebHtmlCli
+- [x]   check: ResizeOverlayImage in webHtmlCli
 - []   check: if (profile.MetaData) // todo: check if works
 - []   check: sync from webUI with '+' in name